cur-TTMS1/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
mysql-connector-python==8.0.33
Werkzeug==2.0.1
python-dotenv==0.19.0 